requirements.txt
flask
apscheduler
groq
feedparser
tweepy
python-dotenv
gunicorn